back-end/DivinaHamburgueriaApi/divinahamburgueriaapi/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.20
pytz==2021.1
requests==2.26.0
rest-condition==1.0.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
